world: Convert EnumError to i128
This fits all enum values, even signed ones.

+++ wow_world_base/src/errors.rs
@@ -4,11 +4,11 @@ use std::fmt::{Display, Formatter};
 #[derive(Debug, Copy, Clone, Ord, PartialOrd, Eq, PartialEq, Hash)]
 pub struct EnumError {
     pub name: &'static str,
-    pub value: u64,
+    pub value: i128,
 }
 
 impl EnumError {
-    pub const fn new(name: &'static str, value: u64) -> Self {
+    pub const fn new(name: &'static str, value: i128) -> Self {
         Self { name, value }
     }
 }
